requirements.txt
# This file is @generated by PDM.
# Please do not edit it manually.

beautifulsoup4==4.12.3
black==24.4.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6; sys_platform == "win32" or platform_system == "Windows"
fastapi
idna==3.7
iniconfig==2.0.0
mypy==1.9.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.2.0
pluggy==1.4.0
pytest==8.1.1
python-dotenv==1.0.1
requests==2.31.0
requests-toolbelt==1.0.0
soupsieve==2.5
typing-extensions==4.9.0
urllib3==2.2.1
